+++ config/config.php
@@ -2,9 +2,41 @@
 
 return [
 
-    'enabled' => true,
+    /*
+    |--------------------------------------------------------------------------
+    | Enable Converter
+    |--------------------------------------------------------------------------
+    |
+    | This option specifies whether the converter feature is on or off.
+    |
+    */
 
-    // Default character replacement
+    'enabled' => true,
+    
+    /*
+    |--------------------------------------------------------------------------
+    | Default Function
+    |--------------------------------------------------------------------------
+    |
+    | This option controls the default function that will be used by the converting.
+    | "strtr" options needs a translate list.
+    | 
+    | Supported: "strtr", "html_entity_decode"
+    |
+    */
+    
+    'default' => 'strtr',
+    
+    /*
+    |--------------------------------------------------------------------------
+    | Translate List
+    |--------------------------------------------------------------------------
+    |
+    | This is translate list for "strtr" function. You can add additional values.
+    | If you use "html_entity_decode" for default function, completely ignore translate list.
+    |
+    */
+    
     'translate' => [
         '&ccedil;' => 'ç',
         '&Ccedil;' => 'Ç',
@@ -13,4 +45,5 @@ return [
         '&uuml;'   => 'ü',
         '&Uuml;'   => 'Ü',
     ]
+
 ];